Reorganize source into lib
