+++ index.ts
@@ -1,5 +1,6 @@
-export { parse } from "./parse-to-unist";
-export { tokenize } from "./tokenize";
-export { compact } from "./unist-compact";
-export { toD3Force } from "./to-d3-force";
-export { toDot } from "./to-dot";
+export { parse } from "./lib/parse-to-unist";
+export { tokenize } from "./lib/tokenize";
+export { compact } from "./lib/unist-compact";
+export { toD3Force } from "./lib/to-d3-force";
+export { toDot } from "./lib/to-dot";
+export { toGraphlib } from "./lib/to-graphlib";
