+++ lib/unist-compact.ts
@@ -1,7 +1,13 @@
 import { Parent } from "unist";
 import visit from "unist-util-visit-parents";
 
+const remove = new Set(["TRLR", "SUBM", "HEAD"]);
+
 export function compact(root: Parent): Parent {
+  // Remove "trailer" objects, which are not useful to us.
+  root.children = root.children.filter((child) => {
+    return !remove.has(child.type);
+  });
   for (let child of root.children) {
     if (!child.data) child.data = {};
     visit(child, (node, ancestors) => {
